more docs in protobuf.*
